images/snapshot-release-tags/requirements.txt
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Deprecated==1.2.13
idna==3.4
pycparser==2.21
PyGithub==1.58.2
PyJWT==2.7.0
PyNaCl==1.5.0
requests==2.30.0
urllib3==2.0.2
wrapt==1.15.0
